severed game state from the main object
